+++ files/default/tests/minitest/apache2_test.rb
@@ -4,7 +4,7 @@ describe_recipe "jira::apache2" do
   include Helpers::Jira
 
   it 'has apache VirtualHost' do
-    file("#{node['apache']['dir']}/sites-available/#{node['stash']['apache2']['virtual_host_name']}.conf").must_exist
+    file("#{node['apache']['dir']}/sites-available/#{node['jira']['apache2']['virtual_host_name']}.conf").must_exist
   end
 
   it 'starts apache' do