+++ test/cookbooks/jira_test/attributes/default.rb
@@ -1,0 +1,10 @@
+
+set["java"]["install_flavor"] = "oracle"
+#set["java"]["java_home"] = "/usr/lib/jvm/java-7-oracle"
+#set["java"]["jdk_version"] = "7"
+set["java"]["oracle"]["accept_oracle_download_terms"] = true
+
+set["mysql"]["bind_address"] = "0.0.0.0"
+set["mysql"]["server_root_password"] = "iloverandompasswords"
+set["mysql"]["server_repl_password"] = "iloverandompasswords"
+set["mysql"]["server_debian_password"] = "iloverandompasswords"