+++ test/cookbooks/jira_test/recipes/default.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: jira
+# Cookbook Name:: jira_test
 # Recipe:: default
 #
-# Copyright 2012
+# Copyright 2013, Brian Flad
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,16 +17,5 @@
 # limitations under the License.
 #
 
-platform = "windows" if node['platform_family'] == "windows"
-platform ||= "linux"
-settings = Jira.settings(node)
-
-include_recipe "jira::database" if settings['database']['host'] == "localhost"
-include_recipe "jira::#{platform}_#{node['jira']['install_type']}"
-
-unless node['jira']['install_type'].match("war")
-  include_recipe "jira::tomcat_configuration"
-  include_recipe "jira::apache2"
-end
-
-include_recipe "jira::configuration"
+include_recipe "java"
+include_recipe "jira"