+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.0.0
+
+* Split default recipe into individual recipes
+* apache2 recipe does not include default recipe
+* Load database/tomcat settings via Jira.settings library (bonus: help support Chef Solo)
+* Moved apache2 attributes into default attributes
+* Bump default JIRA version to 5.2.11
+* Added url_base attribute
+* Auto-detect checksum attribute for some versions
+* Added Vagrantfile and Test Kitchen for testing
+* minitest fixes
+* Added COMPATIBILITY.md
+* Refactored README documentation
+
 ## 0.1.3
 
 * Chef 11 fixes in apache2 recipe